+++ packages/database/src/schemas/device.ts
@@ -1,32 +1,13 @@
-import type { WorkspaceInitResult } from '@lobechat/types';
+import type { WorkingDirEntry } from '@lobechat/types';
 import { index, jsonb, pgTable, text, uniqueIndex, uuid, varchar } from 'drizzle-orm/pg-core';
 
 import { timestamps, timestamptz } from './_helpers';
 import { users } from './user';
 import { workspaces } from './workspace';
 
-/**
- * A working directory the device has used. Structured (rather than a bare path
- * string) so metadata such as the detected repo type survives — a remote client
- * viewing this device can't re-probe its filesystem, so whatever isn't captured
- * here at the source is lost. Mirrors the client-local `RecentDirEntry` shape.
- */
-export interface WorkingDirEntry {
-  path: string;
-  repoType?: 'git' | 'github';
-  /**
-   * Cached "workspace init" scan of this directory (AGENTS.md + project skills).
-   * Populated server-side at run start via `deviceGateway.initWorkspace` and
-   * reused within the TTL gated by `workspaceScannedAt`. Also read directly by
-   * the web UI to render the project's skills / instructions.
-   */
-  workspace?: WorkspaceInitResult;
-  /**
-   * Epoch ms when `workspace` was last scanned. Hoisted to the top level (out of
-   * `workspace`) so freshness can be checked without deserializing the payload.
-   */
-  workspaceScannedAt?: number;
-}
+// `WorkingDirEntry` is defined in `@lobechat/types`; re-exported here so existing
+// `@lobechat/database/schemas` consumers keep importing it from the schema barrel.
+export type { WorkingDirEntry } from '@lobechat/types';
 
 /**
  * Stable device identity anchor — one row per physical machine per user.
