+++ src/features/ChatInput/RuntimeConfig/WorkingDirectoryPicker.tsx
@@ -190,25 +190,23 @@ const AddRemoteFolderRow = memo<{
 }>(({ defaultCwd, deviceId, onBeforeOpen, onPick }) => {
   const { t } = useTranslation('device');
 
-  // Validate the entered path on the target device (it can't be browsed here).
-  // Block only on a definitive negative; an unreachable device (null) is treated
-  // as "can't verify" and allowed through.
-  const validate = async (path: string): Promise<string | undefined> => {
-    if (!deviceId) return undefined;
-    const result = await deviceService.statPath(deviceId, path);
-    if (!result) return undefined;
-    if (!result.exists) return t('workingDirectory.pathNotExist');
-    if (!result.isDirectory) return t('workingDirectory.pathNotDirectory');
+  // Stat the entered path on the target device (it can't be browsed here): block
+  // on a definitive negative, otherwise commit with the detected repoType so the
+  // recent entry shows the right (git / github) icon. An unreachable device
+  // (null) is treated as "can't verify" and allowed through without a repoType.
+  const handleSubmit = async (path: string): Promise<string | undefined> => {
+    const result = deviceId ? await deviceService.statPath(deviceId, path) : undefined;
+    if (result) {
+      if (!result.exists) return t('workingDirectory.pathNotExist');
+      if (!result.isDirectory) return t('workingDirectory.pathNotDirectory');
+    }
+    onPick({ path, repoType: result?.repoType });
     return undefined;
   };
 
   const handleClick = () => {
     onBeforeOpen();
-    openAddWorkingDirModal({
-      onSubmit: (path) => onPick({ path }),
-      placeholder: defaultCwd || undefined,
-      validate,
-    });
+    openAddWorkingDirModal({ onSubmit: handleSubmit, placeholder: defaultCwd || undefined });
   };
   return (
     <Flexbox
