+++ src/store/device/initialState.ts
@@ -1,11 +1,6 @@
-import type { deviceService } from '@/services/device';
+import type { DeviceListItem } from '@lobechat/types';
 
-/**
- * A device row as returned by `device.listDevices` — either a registered device
- * or an online-only "ghost" (connected but not yet persisted). Inferred from the
- * service so the store stays in sync with the contract.
- */
-export type DeviceListItem = Awaited<ReturnType<typeof deviceService.listDevices>>[number];
+export type { DeviceListItem };
 
 export interface DeviceState {
   devices: DeviceListItem[];
